requirements.txt
groq
edge-tts
moviepy==1.0.3
Pillow==9.5.0
requests
